refactor: Consolidate split backend architecture into Express
