+++ backend/app.js
@@ -4,6 +4,8 @@ import cors from "cors";
 import chatRoutes from "./routers/chatRoutes.js";
 import aiRoutes from "./routers/aiRoutes.js";
 import matchRoutes from "./routers/matchRoutes.js";
+import cronRoutes from "./routers/cronRoutes.js";
+import notificationRoutes from "./routers/notificationRoutes.js";
 import { errorHandler } from "./middlewares/errorHandler.js";
 
 const app = express();
@@ -31,6 +33,8 @@ app.get("/health", (_req, res) => {
 app.use("/api/ai", aiRoutes);
 app.use("/api", chatRoutes);
 app.use("/api/match", matchRoutes);
+app.use("/api/cron", cronRoutes);
+app.use("/api/notifications", notificationRoutes);
 
 // 404 handler for unmatched routes
 app.use((_req, res) => {
