--- supabase/functions/send-push-notification/index.ts
@@ -1,96 +1,0 @@
-import { serve } from "https://deno.land/std@0.224.0/http/server.ts";
-import { createClient } from "npm:@supabase/supabase-js@2";
-import webpush from "npm:web-push@3.6.7";
-
-serve(async (req) => {
-  if (req.method !== "POST") {
-    return new Response(JSON.stringify({ error: "Method not allowed" }), {
-      status: 405,
-      headers: { "Content-Type": "application/json" },
-    });
-  }
-
-  const supabaseUrl = Deno.env.get("SUPABASE_URL");
-  const serviceRoleKey = Deno.env.get("SUPABASE_SERVICE_ROLE_KEY");
-  const vapidPublicKey = Deno.env.get("VAPID_PUBLIC_KEY");
-  const vapidPrivateKey = Deno.env.get("VAPID_PRIVATE_KEY");
-  const vapidSubject = Deno.env.get("VAPID_SUBJECT") || "mailto:admin@example.com";
-
-  if (!supabaseUrl || !serviceRoleKey || !vapidPublicKey || !vapidPrivateKey) {
-    return new Response(JSON.stringify({ error: "Missing push server env" }), {
-      status: 500,
-      headers: { "Content-Type": "application/json" },
-    });
-  }
-
-  const authHeader = req.headers.get("Authorization");
-  const webhookSecret = Deno.env.get("WEBHOOK_SECRET");
-  let isAuthorized = false;
-
-  if (webhookSecret && authHeader === `Bearer ${webhookSecret}`) {
-    isAuthorized = true;
-  } else if (authHeader) {
-    const supabaseAuth = createClient(supabaseUrl, Deno.env.get("SUPABASE_ANON_KEY") || serviceRoleKey);
-    const { data: { user } } = await supabaseAuth.auth.getUser(authHeader.replace("Bearer ", ""));
-    if (user) isAuthorized = true;
-  }
-
-  if (!isAuthorized) {
-    return new Response(JSON.stringify({ error: "Unauthorized" }), {
-      status: 401,
-      headers: { "Content-Type": "application/json" },
-    });
-  }
-
-  const { user_id, title, body, action_url } = await req.json();
-
-  if (!user_id || !title || !body) {
-    return new Response(JSON.stringify({ error: "user_id, title, and body are required" }), {
-      status: 400,
-      headers: { "Content-Type": "application/json" },
-    });
-  }
-
-  webpush.setVapidDetails(vapidSubject, vapidPublicKey, vapidPrivateKey);
-
-  const supabase = createClient(supabaseUrl, serviceRoleKey);
-
-  const { data: subscriptions, error } = await supabase
-    .from("push_subscriptions")
-    .select("id,endpoint,p256dh,auth")
-    .eq("user_id", user_id);
-
-  if (error) {
-    return new Response(JSON.stringify({ error: error.message }), {
-      status: 500,
-      headers: { "Content-Type": "application/json" },
-    });
-  }
-
-  const results = await Promise.allSettled(
-    (subscriptions || []).map((subscription) =>
-      webpush.sendNotification(
-        {
-          endpoint: subscription.endpoint,
-          keys: {
-            p256dh: subscription.p256dh,
-            auth: subscription.auth,
-          },
-        },
-        JSON.stringify({
-          title,
-          body,
-          action_url: action_url || "/notifications",
-        })
-      )
-    )
-  );
-
-  return new Response(
-    JSON.stringify({
-      sent: results.filter((result) => result.status === "fulfilled").length,
-      failed: results.filter((result) => result.status === "rejected").length,
-    }),
-    { headers: { "Content-Type": "application/json" } }
-  );
-});
